+++ buddypress/members/single/messages.php
@@ -6,8 +6,12 @@
  * @subpackage bp-legacy
  */
 
-?>
+the_telabotanica_module('header-dashboard', [
+	'title' => __('Mes messages', 'telabotanica')
+]);
 
+?>
+<div id="buddypress">
 <div class="item-list-tabs no-ajax" id="subnav" aria-label="<?php esc_attr_e( 'Member secondary navigation', 'buddypress' ); ?>" role="navigation">
 	<ul>
 
@@ -107,3 +111,5 @@ switch ( bp_current_action() ) :
 		bp_get_template_part( 'members/single/plugins' );
 		break;
 endswitch;
+?>
+</div>