+++ modules/title/title.php
@@ -3,11 +3,25 @@
 	$defaults = [
 		'title' => 'Titre',
 		'level' => 1,
+		'suffix' => false,
+		'href' => false,
 		'modifiers' => []
 	];
 
 	$data = telabotanica_styleguide_data($defaults, $data);
 	$data->modifiers = telabotanica_styleguide_modifiers_array('title', $data->modifiers);
+
+	if ( $data->suffix ) :
+		$data->title .= sprintf( '<span class="title-suffix">%s</span>', $data->suffix );
+	endif;
+
+	if ( $data->href ) :
+		$data->title = sprintf(
+			'<a href="%s">%s</a>',
+			$data->href,
+			$data->title
+		);
+	endif;
 
 	printf(
 		'<h%s class="%s">%s</h%s>',