+++ CHANGELOG.md
@@ -11,7 +11,20 @@ plus de détails sur http://semver.org/lang/fr/
 
 ## x.β.α (bientôt)
 
+* composant `map` : centre optionnel
+* module `block-dashboard` ajouté
+* module `button` : ajout de la couleur rouge
+* module `cover-home` : ajout du lien vers le tableau de bord
+* module `cover-member` ajouté
+* module `feed` ajouté
+* module `feed-date` ajouté
+* module `feed-item` ajouté
+* module `header` : fix `z-index`
+* module `header-dashboard` ajouté
 * module `notice` ajouté
+* module `title` : support d'un suffixe optionnel
+* layout `2-col` : ajout d'une version `larger-first-col`
+* layout `content-col` : ajout d'une largeur max en version `is-dashboard`
 
 ## 0.0.7 (2017-03-04)